+++ cli/cmd/install_test.go
@@ -37,7 +37,7 @@ func TestRender(t *testing.T) {
 		ControllerLogLevel:               "ControllerLogLevel",
 		ControllerComponentLabel:         "ControllerComponentLabel",
 		CreatedByAnnotation:              "CreatedByAnnotation",
-		ProxyAPIPort:                     123,
+		DestinationAPIPort:               123,
 		EnableTLS:                        true,
 		TLSTrustAnchorVolumeName:         "TLSTrustAnchorVolumeName",
 		TLSSecretsVolumeName:             "TLSSecretsVolumeName",
@@ -88,7 +88,7 @@ func TestRender(t *testing.T) {
 		ControllerLogLevel:               "ControllerLogLevel",
 		ControllerComponentLabel:         "ControllerComponentLabel",
 		CreatedByAnnotation:              "CreatedByAnnotation",
-		ProxyAPIPort:                     123,
+		DestinationAPIPort:               123,
 		ProxyUID:                         2102,
 		ControllerUID:                    2103,
 		EnableTLS:                        true,
