+++ cli/cmd/testdata/inject_emojivoto_list.golden.yml
@@ -43,7 +43,7 @@ items:
           - name: LINKERD2_PROXY_LOG
             value: warn,linkerd2_proxy=info
           - name: LINKERD2_PROXY_CONTROL_URL
-            value: tcp://linkerd-proxy-api.linkerd.svc.cluster.local:8086
+            value: tcp://linkerd-destination.linkerd.svc.cluster.local:8086
           - name: LINKERD2_PROXY_CONTROL_LISTENER
             value: tcp://0.0.0.0:4190
           - name: LINKERD2_PROXY_METRICS_LISTENER
@@ -147,7 +147,7 @@ items:
           - name: LINKERD2_PROXY_LOG
             value: warn,linkerd2_proxy=info
           - name: LINKERD2_PROXY_CONTROL_URL
-            value: tcp://linkerd-proxy-api.linkerd.svc.cluster.local:8086
+            value: tcp://linkerd-destination.linkerd.svc.cluster.local:8086
           - name: LINKERD2_PROXY_CONTROL_LISTENER
             value: tcp://0.0.0.0:4190
           - name: LINKERD2_PROXY_METRICS_LISTENER
