+++ cli/cmd/testdata/inject_emojivoto_pod_tls.golden.yml
@@ -24,7 +24,7 @@ spec:
     - name: LINKERD2_PROXY_LOG
       value: warn,linkerd2_proxy=info
     - name: LINKERD2_PROXY_CONTROL_URL
-      value: tcp://linkerd-proxy-api.linkerd.svc.cluster.local:8086
+      value: tcp://linkerd-destination.linkerd.svc.cluster.local:8086
     - name: LINKERD2_PROXY_CONTROL_LISTENER
       value: tcp://0.0.0.0:4190
     - name: LINKERD2_PROXY_METRICS_LISTENER
