+++ controller/api/destination/endpoints_watcher.go
@@ -1,4 +1,4 @@
-package proxy
+package destination
 
 import (
 	"fmt"
