+++ controller/api/destination/k8s_resolver.go
@@ -1,4 +1,4 @@
-package proxy
+package destination
 
 import (
 	"errors"
