+++ controller/api/destination/k8s_resolver_test.go
@@ -1,4 +1,4 @@
-package proxy
+package destination
 
 import (
 	"fmt"
