+++ controller/api/destination/profile_watcher_test.go
@@ -1,4 +1,4 @@
-package proxy
+package destination
 
 import (
 	"reflect"
