+++ controller/cmd/public-api/main.go
@@ -25,7 +25,7 @@ func main() {
 	kubeConfigPath := flag.String("kubeconfig", "", "path to kube config")
 	prometheusURL := flag.String("prometheus-url", "http://127.0.0.1:9090", "prometheus url")
 	metricsAddr := flag.String("metrics-addr", ":9995", "address to serve scrapable metrics on")
-	proxyAPIAddr := flag.String("proxy-api-addr", "127.0.0.1:8086", "address of proxy-api service")
+	destinationAPIAddr := flag.String("destination-addr", "127.0.0.1:8086", "address of destination service")
 	tapAddr := flag.String("tap-addr", "127.0.0.1:8088", "address of tap service")
 	controllerNamespace := flag.String("controller-namespace", "linkerd", "namespace in which Linkerd is installed")
 	singleNamespace := flag.Bool("single-namespace", false, "only operate in the controller namespace")
@@ -41,12 +41,12 @@ func main() {
 	}
 	defer tapConn.Close()
 
-	proxyAPIConn, err := grpc.Dial(*proxyAPIAddr, grpc.WithInsecure())
+	destinationAPIConn, err := grpc.Dial(*destinationAPIAddr, grpc.WithInsecure())
 	if err != nil {
 		log.Fatal(err.Error())
 	}
-	defer proxyAPIConn.Close()
-	discoveryClient := discovery.NewDiscoveryClient(proxyAPIConn)
+	defer destinationAPIConn.Close()
+	discoveryClient := discovery.NewDiscoveryClient(destinationAPIConn)
 
 	k8sClient, err := k8s.NewClientSet(*kubeConfigPath)
 	if err != nil {
